main2.py
import numpy as np
import torch
import torch.nn as nn
import torch.optim as optim
from torch.utils.data import DataLoader, TensorDataset
import matplotlib.pyplot as plt
from scipy.ndimage import convolve

# Simple 2D scalar field theory simulator
class ScalarFieldTheory:
    def __init__(self, L=32, J=1.0):
        """
        L: lattice size (L x L)
        J: coupling strength
        """
        self.L = L
        self.J = J
        
    def energy(self, config):
        """Compute energy of a field configuration"""
        # Interaction term: neighbors want to align
        kernel = np.array([[0, 1, 0], [1, 0, 1], [0, 1, 0]])
        neighbor_sum = convolve(config, kernel, mode='wrap')
        interaction = -self.J * np.sum(config * neighbor_sum) / 2
        
        # Field strength term
        field_term = np.sum(config**2) / 2
        
        return interaction + field_term
    
    def monte_carlo_step(self, config, beta):
        """One Monte Carlo update sweep"""
        for _ in range(self.L**2):
            # Random site
            i, j = np.random.randint(0, self.L, 2)
            
            # Compute local energy
            neighbors = (
                config[(i+1)%self.L, j] + config[(i-1)%self.L, j] +
                config[i, (j+1)%self.L] + config[i, (j-1)%self.L]
            )
            
            # Propose new value
            old_val = config[i, j]
            new_val = old_val + np.random.normal(0, 0.5)
            
            # Compute energy change
            delta_E = (new_val**2 - old_val**2) / 2
            delta_E -= self.J * (new_val - old_val) * neighbors
            
            # Accept/reject
            if delta_E < 0 or np.random.random() < np.exp(-beta * delta_E):
                config[i, j] = new_val
                
        return config
    
    def generate_configuration(self, beta, n_steps=500):
        """Generate equilibrium configuration at temperature 1/beta"""
        config = np.random.randn(self.L, self.L) * 0.1
        
        # Equilibration
        for _ in range(n_steps):
            config = self.monte_carlo_step(config, beta)
            
        return config

# Neural network for phase classification
class PhaseClassifier(nn.Module):
    def __init__(self, L=32):
        super().__init__()
        self.conv1 = nn.Conv2d(1, 16, kernel_size=3, padding=1)
        self.conv2 = nn.Conv2d(16, 32, kernel_size=3, padding=1)
        self.conv3 = nn.Conv2d(32, 64, kernel_size=3, padding=1)
        self.pool = nn.MaxPool2d(2)
        
        # Calculate size after convolutions and pooling
        size = L // 8  # Three pooling layers
        self.fc1 = nn.Linear(64 * size * size, 128)
        self.fc2 = nn.Linear(128, 2)
        
        self.relu = nn.ReLU()
        self.dropout = nn.Dropout(0.5)
        
    def forward(self, x):
        x = self.pool(self.relu(self.conv1(x)))
        x = self.pool(self.relu(self.conv2(x)))
        x = self.pool(self.relu(self.conv3(x)))
        x = x.view(x.size(0), -1)
        x = self.relu(self.fc1(x))
        x = self.dropout(x)
        x = self.fc2(x)
        return x

# Generate training data
def generate_dataset(n_samples=1000, L=32):
    """Generate labeled configurations from known phases"""
    sft = ScalarFieldTheory(L=L)
    
    # High temperature (disordered) - beta = 0.1 to 0.3
    disordered_configs = []
    for _ in range(n_samples // 2):
        beta = np.random.uniform(0.1, 0.3)
        config = sft.generate_configuration(beta)
        disordered_configs.append(config)
    
    # Low temperature (ordered) - beta = 0.7 to 1.0  
    ordered_configs = []
    for _ in range(n_samples // 2):
        beta = np.random.uniform(0.7, 1.0)
        config = sft.generate_configuration(beta)
        ordered_configs.append(config)
    
    # Combine and create labels
    configs = np.array(disordered_configs + ordered_configs)
    labels = np.array([0] * (n_samples // 2) + [1] * (n_samples // 2))
    
    # Shuffle
    idx = np.random.permutation(n_samples)
    configs = configs[idx]
    labels = labels[idx]
    
    return configs, labels

# Training function
def train_classifier(model, train_loader, val_loader, epochs=15):
    criterion = nn.CrossEntropyLoss()
    optimizer = optim.Adam(model.parameters(), lr=0.001)
    
    train_losses = []
    val_accuracies = []
    
    for epoch in range(epochs):
        # Training
        model.train()
        total_loss = 0
        for configs, labels in train_loader:
            optimizer.zero_grad()
            outputs = model(configs)
            loss = criterion(outputs, labels)
            loss.backward()
            optimizer.step()
            total_loss += loss.item()
        
        train_losses.append(total_loss / len(train_loader))
        
        # Validation
        model.eval()
        correct = 0
        total = 0
        with torch.no_grad():
            for configs, labels in val_loader:
                outputs = model(configs)
                _, predicted = torch.max(outputs, 1)
                total += labels.size(0)
                correct += (predicted == labels).sum().item()
        
        accuracy = correct / total
        val_accuracies.append(accuracy)
        
        if (epoch + 1) % 5 == 0:
            print(f'Epoch {epoch+1}: Loss = {train_losses[-1]:.4f}, '
                  f'Validation Accuracy = {accuracy:.4f}')
    
    return train_losses, val_accuracies

# Detect phase transition
def detect_phase_transition(model, beta_range, n_samples=50):
    """Use trained model to find the phase transition"""
    sft = ScalarFieldTheory(L=32)
    predictions = []
    
    model.eval()
    with torch.no_grad():
        for beta in beta_range:
            ordered_count = 0
            for _ in range(n_samples):
                config = sft.generate_configuration(beta)
                config_tensor = torch.FloatTensor(config).unsqueeze(0).unsqueeze(0)
                output = model(config_tensor)
                _, predicted = torch.max(output, 1)
                if predicted.item() == 1:  # Ordered phase
                    ordered_count += 1
            
            predictions.append(ordered_count / n_samples)
    
    return np.array(predictions)

# Main demonstration
if __name__ == "__main__":
    print("AI Discovers Phase Transitions in Quantum Field Theory")
    print("=" * 50)
    
    # Set random seeds for reproducibility
    np.random.seed(42)
    torch.manual_seed(42)
    
    # Generate data
    print("\n1. Generating training data from known phases...")
    train_configs, train_labels = generate_dataset(n_samples=1000)
    val_configs, val_labels = generate_dataset(n_samples=200)
    
    # Convert to PyTorch tensors
    train_configs = torch.FloatTensor(train_configs).unsqueeze(1)
    train_labels = torch.LongTensor(train_labels)
    val_configs = torch.FloatTensor(val_configs).unsqueeze(1)
    val_labels = torch.LongTensor(val_labels)
    
    # Create data loaders
    train_dataset = TensorDataset(train_configs, train_labels)
    val_dataset = TensorDataset(val_configs, val_labels)
    train_loader = DataLoader(train_dataset, batch_size=32, shuffle=True)
    val_loader = DataLoader(val_dataset, batch_size=32)
    
    # Train model
    print("\n2. Training neural network to recognize phases...")
    model = PhaseClassifier(L=32)
    train_losses, val_accuracies = train_classifier(
        model, train_loader, val_loader, epochs=15
    )
    
    # Detect phase transition
    print("\n3. Using AI to find the phase transition...")
    beta_range = np.linspace(0.1, 1.0, 30)
    predictions = detect_phase_transition(model, beta_range)
    
    # Find transition point (where predictions = 0.5)
    transition_idx = np.argmin(np.abs(predictions - 0.5))
    beta_critical = beta_range[transition_idx]
    print(f"\nAI-detected critical temperature: β_c ≈ {beta_critical:.3f}")
    print(f"(Theoretical value: β_c ≈ 0.44)")
    
    # Visualize results
    fig, axes = plt.subplots(2, 2, figsize=(12, 10))
    
    # Sample configurations
    sft = ScalarFieldTheory(L=32)
    disordered = sft.generate_configuration(beta=0.2)
    ordered = sft.generate_configuration(beta=0.8)
    
    axes[0, 0].imshow(disordered, cmap='coolwarm', vmin=-3, vmax=3)
    axes[0, 0].set_title('Disordered Phase (High T)')
    axes[0, 0].axis('off')
    
    axes[0, 1].imshow(ordered, cmap='coolwarm', vmin=-3, vmax=3)
    axes[0, 1].set_title('Ordered Phase (Low T)')
    axes[0, 1].axis('off')
    
    # Training history
    axes[1, 0].plot(train_losses)
    axes[1, 0].set_xlabel('Epoch')
    axes[1, 0].set_ylabel('Training Loss')
    axes[1, 0].set_title('Learning Progress')
    axes[1, 0].grid(True)
    
    # Phase transition detection
    axes[1, 1].plot(beta_range, predictions, 'b-', linewidth=2)
    axes[1, 1].axhline(y=0.5, color='r', linestyle='--', label='Transition')
    axes[1, 1].axvline(x=beta_critical, color='g', linestyle='--', 
                       label=f'AI: β_c={beta_critical:.3f}')
    axes[1, 1].set_xlabel('β (inverse temperature)')
    axes[1, 1].set_ylabel('Probability of Ordered Phase')
    axes[1, 1].set_title('AI-Detected Phase Transition')
    axes[1, 1].legend()
    axes[1, 1].grid(True)
    
    plt.tight_layout()
    plt.savefig('phase_transition_detection.png', dpi=150)
    plt.show()
    
    print("\n" + "="*50)
    print("Success! The neural network has learned to recognize phase transitions")
    print("without being taught any physics. This same approach scales to")
    print("complex quantum field theories where analytical methods fail.")