requirements.txt
python >= 3.8
pytorch >= 1.10
numpy
scipy
matplotlib